# hackathon-backend/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Hackathon, VMO-Global Car Sharing
  description: 'VMO-Global Car Sharing Api Docs'
  version: 1.0.0
servers:
- url: http://localhost:8082/api/v1/
...
paths:
...
  /customer:
    get:
      tags:
      - Customer
      summary: List of customers with status is actived
      description: list customers
      operationId: geListCustomers
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                  data: 
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
        500:
          description: "Internal Network Error"
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "UNKNOWN_TRANSACTION_STATUS"
                  message: 
                    type: string
                    description: "UNAVAILABLE: io exception"
    post:
      tags:
      - Customer
      summary: Add a new customer
      operationId: addCustomer
      requestBody:
        description: Customer object that needs to be added to the store
        content:
          application/json:
            schema:
              type: object
              properties: 
                firstName: 
                  type: string
                lastName: 
                  type: string
                address: 
                  type: string
                mobilephone: 
                  type: string
                createBy: 
                    type: string
        required: true
      responses:
        200:
          description: "Successfully"
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                  message: 
                    type: string
        500:
          description: "Internal Network Error"
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "UNKNOWN_TRANSACTION_STATUS"
                  message: 
                    type: string
                    description: "UNAVAILABLE: io exception"
  /customer/{id}:
    get:
      tags:
      - Customer
      summary: Find details customer by ID
      description: Returns details customer
      operationId: getCustomerById
      parameters:
      - name: id
        in: path
        description: ID of customer to return
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                  data: 
                    $ref: '#/components/schemas/Customer'
        500:
          description: Internal Server Error
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "CONTRACT_CONTEXTUAL_ERROR"
                  message: 
                    type: string
                    description: "Customer does not exist!"
    put:
      tags:
      - Customer
      summary: Updates a customer in the store with form data
      operationId: updateCustomer
      parameters:
      - name: id
        in: path
        description: ID of customer that needs to be updated
        required: true
        schema:
          type: string
      requestBody:
        content:
            application/json:
              schema:
                type: object
                properties: 
                  firstName: 
                    type: string
                  lastName: 
                    type: string
                  address: 
                    type: string
                  mobilephone: 
                    type: string
                  updateBy: 
                    type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "Succeeded!"
                  message: 
                    type: string
                    description: "Customer update successfully!"
        500:
          description: Internal Server Error
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "CONTRACT_CONTEXTUAL_ERROR"
                  message: 
                    type: string
                    description: "Customer does not exist!"
...
  /customer/all:
    get:
      tags:
      - Customer
      summary: List of all customers ( includes actived and inactived )
      description: list customers
      operationId: geListALlCustomers
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                  data: 
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
        500:
          description: "Internal Network Error"
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "UNKNOWN_TRANSACTION_STATUS"
                  message: 
                    type: string
                    description: "UNAVAILABLE: io exception"
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: string
        firstName: 
          type: string
        lastName:
          type: string
        address:
          type: string
        mobilephone:
          type: string
        isDeleted:
          type: number
          format: int1
          default: 0
        isActive:
          type: number
          format: int1
          default: 1
        createBy:
          type: string
        updateBy:
          type: string
        createAt:
          type: string
        updateAt:
          type: string
...